requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
ollama
requests
tqdm
